HHH-18585 move AfterUseAction and BeforeUseAction to an SPI package
