+++ hibernate-core/src/main/java/org/hibernate/query/sqm/mutation/spi/BeforeUseAction.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.query.sqm.mutation.internal.temptable;
+package org.hibernate.query.sqm.mutation.spi;
 
 /**
  * Actions to perform in regards to a temporary table prior to each use.
